requirements.txt
Flask
Flask-PyMongo
python-dotenv
flask_cors
flask_jwt_extended